circular imports in getters fixed and getorbuild_cell_type_dict() working
